Fixing Postgres DataStore Tests
