+++ document-store/src/main/java/org/hypertrace/core/documentstore/postgres/PostgresDatastore.java
@@ -115,8 +115,7 @@ public class PostgresDatastore implements Datastore {
     return false;
   }
   
-  @VisibleForTesting
-  Connection getPostgresClient() {
+  public Connection getPostgresClient() {
     return client;
   }
 }
